requirements.txt
arrow==1.2.3
asgiref==3.6.0
attrs==23.1.0
blessed==1.20.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-picklefield==3.1
django-q==1.3.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
PyJWT==2.6.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
redis==3.5.3
requests==2.28.2
six==1.16.0
sqlparse==0.4.4
stripe==5.4.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
wcwidth==0.2.6
